Mysite/requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.1
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
configparser==3.5.0
Django==1.11.14
django-ckeditor==5.6.1
django-contrib-comments==1.8.0
django-js-asset==1.1.0
django-modelcluster==4.1
django-taggit==0.22.2
django-treebeard==4.3
django-wysiwyg-redactor==0.5.1
djangorestframework==3.8.2
draftjs-exporter==2.1.2
enum34==1.1.6
filebrowser-safe==0.5.0
future==0.16.0
futures==3.2.0
grappelli-safe==0.5.0
html5lib==1.0.1
idna==2.7
isort==4.3.4
l18n==2016.6.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Mezzanine==4.3.0
oauthlib==2.1.0
Pillow==5.2.0
pylint==1.9.2
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
singledispatch==3.4.0.3
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21
urllib3==1.23
webencodings==0.5.1
Willow==1.1
wrapt==1.10.11
